Moved Stuff around and fixed trigger job
