+++ Assets/Scripts/Systems/Execution/AntExecutionSystem.cs
@@ -41,6 +41,8 @@ namespace Systems
                         mass.GetImpulseFromForce(dir, ForceMode.Force, in deltaTime, out var impulse, out var massImpulse);
                         vel.ApplyLinearImpulse(in massImpulse, in impulse);
                         vel.ApplyAngularImpulse(in mass, 5f * deltaTime * math.up() * meth.SignedAngle(ltw.Forward, dir, math.up()));
+
+                        state.executionLine--;
                         break;
                     default: return;
                 }
