+++ Assets/Scripts/Systems/Execution/ExecutionSystemGroup.cs
@@ -1,0 +1,15 @@
+﻿using Unity.Entities;
+
+namespace Systems.Execution
+{
+    [UnityEngine.ExecuteAlways]
+    [UpdateInGroup(typeof(FixedStepSimulationSystemGroup))]
+    public class ExecutionSystemGroup : ComponentSystemGroup
+    {
+        protected override void OnUpdate()
+        {
+            base.OnUpdate();
+            base.OnUpdate();
+        }
+    }
+}
